+++ src/Ar/ATN/Director.cpp
@@ -2,9 +2,9 @@
  * File: Director.cpp
  * Copyright (c) 2023 Loupe
  * https://loupe.team
- * 
+ *
  * This file is part of All Together Now - ATN, licensed under the MIT License.
- * 
+ *
  */
 
 #include <iostream>
@@ -22,24 +22,6 @@ Director::~Director()
 
 }
 
-unsigned int Director::countActiveThreads(){
-    return threads.size();
-}
-
-void Director::addBehavior( const std::string action,  AtnAPI_typ* api, void *_pParameters, size_t _sParameters ){
-
-    auto it = actions.find(action);
-
-    if (it != actions.end()){
-        it->second.subscribe( api,_pParameters, _sParameters);
-    }
-    else{
-        Action newAction(action);
-        newAction.subscribe(api,_pParameters, _sParameters);
-        actions.insert( std::pair<std::string, Action>(action, newAction));        
-    }
-}
-
 void Director::addState( const std::string state,  AtnAPIState_typ* api, void *_pParameters, size_t _sParameters ){
 
     auto it = states.find(state);
@@ -50,7 +32,7 @@ void Director::addState( const std::string state,  AtnAPIState_typ* api, void *_
     else{
         State newAction( state );
         newAction.subscribe(api, _pParameters, _sParameters);
-        states.insert( std::pair<std::string, State>(state, newAction));        
+        states.insert( std::pair<std::string, State>(state, newAction));
     }
 }
 
@@ -65,10 +47,10 @@ void Director::addState( const std::string state, const std::string name, char *
 	else{
 		State newAction( state );
 		newAction.subscribe( name, value, moduleByPass, moduleStatus, _pParameters, _sParameters);
-		states.insert( std::pair<std::string, State>(state, newAction));        
+		states.insert( std::pair<std::string, State>(state, newAction));
 	}
 
-	
+
 }
 
 void Director::addStateBool( const std::string state, const std::string name, bool* value ){
@@ -81,7 +63,7 @@ void Director::addStateBool( const std::string state, const std::string name, bo
 	else{
 		State newAction( state );
 		newAction.subscribe( name, value);
-		states.insert( std::pair<std::string, State>(state, newAction));        
+		states.insert( std::pair<std::string, State>(state, newAction));
 	}
 }
 
@@ -95,10 +77,10 @@ void Director::addResourceBool( const std::string state, const std::string name,
 	else{
 		State newAction( state );
 		newAction.subscribe( name, pResourceUid, value);
-		states.insert( std::pair<std::string, State>(state, newAction));        
+		states.insert( std::pair<std::string, State>(state, newAction));
 	}
 }
-		
+
 
 
 void Director::addStateBool( const std::string state, const std::string name, bool* value,  void *_pParameters, size_t _sParameters){
@@ -111,7 +93,7 @@ void Director::addStateBool( const std::string state, const std::string name, bo
 	else{
 		State newAction( state );
 		newAction.subscribe( name, value, _pParameters, _sParameters);
-		states.insert( std::pair<std::string, State>(state, newAction));        
+		states.insert( std::pair<std::string, State>(state, newAction));
 	}
 }
 
@@ -125,7 +107,7 @@ void Director::addCommandBool( const std::string command, const std::string modu
 	else{
 		State newAction( command );
 		newAction.subscribe( moduleName, check);
-		commands.insert( std::pair<std::string, State>(command, newAction));        
+		commands.insert( std::pair<std::string, State>(command, newAction));
 	}
 }
 
@@ -139,7 +121,7 @@ void Director::addCommandPLCOpen( const std::string command, const std::string m
 	else{
 		State newAction( command );
 		newAction.subscribe( moduleName, commandBit, status);
-		commands.insert( std::pair<std::string, State>(command, newAction));        
+		commands.insert( std::pair<std::string, State>(command, newAction));
 	}
 }
 
@@ -153,26 +135,9 @@ void Director::addCommandPLCOpen( const std::string command, const std::string m
 	else{
 		State newAction( command );
 		newAction.subscribe( moduleName, commandBit, status, _pParameters, _sParameters);
-		commands.insert( std::pair<std::string, State>(command, newAction));        
+		commands.insert( std::pair<std::string, State>(command, newAction));
 	}
 }
-
-void Director::executeAction( const std::string action,  AtnApiStatus_typ* _pStatus, void *_pParameters, size_t _sParameters){
-
-    auto it = actions.find(action);
-
-    if (it != actions.end()){
-        threads.push_back(it->second); 
-        ///TODO: check if it's OK to start
-        threads.back().start( _pStatus, _pParameters, _sParameters );
-    }
-    else{
-		if( this->outstream ){
-			*this->outstream << "Action not found\n";
-		}
-    }
-}
-
 
 bool Director::executeCommand( const std::string command ){
 
@@ -184,7 +149,7 @@ bool Director::executeCommand( const std::string command ){
     }
     else{
 		if( this->outstream ){
-			*this->outstream << "Action not found\n";
+			*this->outstream << "Command not found\n";
 		}
 		return false;
 	}
@@ -199,7 +164,7 @@ void Director::resetCommand( const std::string command ){
     }
     else{
 		if( this->outstream ){
-			*this->outstream << "Action not found\n";
+			*this->outstream << "Command not found\n";
 		}
     }
 }
@@ -208,7 +173,7 @@ State * Director::getState( const std::string state ){
     auto it = states.find(state);
 
     if (it != states.end()){
-        return &it->second;        
+        return &it->second;
     }
     else{
         return 0;
@@ -219,37 +184,13 @@ State * Director::getCommand( const std::string cmd ){
     auto it = commands.find(cmd);
 
     if (it != commands.end()){
-        return &it->second;        
+        return &it->second;
     }
     else{
         return 0;
     }
 }
 
-void Director::cyclic(){
-
-    for (auto thread = threads.begin(); thread != threads.end(); ++thread){
-        //Run thread
-        if( thread->update() ){
-            thread->print( *this->outstream );
-            //If it's done, remove it            
-            threads.erase( thread );
-        }
-    }
-}
-
-void Director::printState( std::ostream &out ){
-    for (auto thread = threads.begin(); thread != threads.end(); ++thread){
-        thread->print( out );   
-    }
-}
-
-void Director::printActions( std::ostream &out ){
-    out << "\nActions:" << "\n";
-    for( auto action : actions ){
-        out << action.first << "\n";
-    }
-}
 void Director::printStates( std::ostream &out ){
     out << "\nStates:" << "\n";
     for( auto state : states ){
@@ -272,7 +213,7 @@ void Director::printSystemJson( std::ostream &out ){
 	out << "\"commands\":[";
 	comma = 0;
 	for( auto command : commands ){
-		if(comma) out << ","; 
+		if(comma) out << ",";
 		out << "\"" <<command.first << "\"";
 		comma = 1;
 	}
@@ -280,10 +221,10 @@ void Director::printSystemJson( std::ostream &out ){
 	out << ",\"States\":[";
 	comma = 0;
 	for( auto state : states ){
-		if(comma) out << ","; 
+		if(comma) out << ",";
 		out << "\"" <<state.first << "\"";
 		comma = 1;
 	}
 	out << "]";
-	out << "}";	
+	out << "}";
 }