+++ src/Ar/ATN/includes/atnApi.h
@@ -2,9 +2,9 @@
  * File: atnApi.h
  * Copyright (c) 2023 Loupe
  * https://loupe.team
- * 
+ *
  * This file is part of All Together Now - ATN, licensed under the MIT License.
- * 
+ *
  */
 
 #include "atn.h"
@@ -14,33 +14,9 @@
 	{
 #endif
 
-//void atncyclic( void * director );
-
-//Register a behavior to an acc
-void registerBehavior( const STRING *action, const STRING *moduleName, AtnAPI_typ *behavior, UDINT *_pParameters, UDINT _sParameters);
-
-void executeAction( const STRING *action );
-void executeActionReport( const STRING *action, AtnApiStatusLocal_typ *api);
-
-bool oneShot( AtnAPI_typ *Behavior );
-bool oneShotReset( AtnAPI_typ *Behavior, bool *cmd );
-bool oneShotStatus( AtnAPI_typ *Behavior, STRING *status);
-// bool oneShot( AtnAPI_typ *Behavior, bool *cmd, STRING *status );
-ATN_ST_enum respond( AtnAPI_typ *Behavior );
-void readCallState( AtnApiStatusLocal_typ *status);
-
 void atnSetDirector( void *director );
 
-
-
-//bool registerState( STRING *state, STRING *moduleName, AtnAPIState_typ *api);
-//bool registerStateWithParameters( STRING *state, STRING *moduleName, AtnAPIState_typ *api, UDINT * pParameters, UDINT sParameters);
-
-//bool subscribeCommandBool(plcstring* state, plcstring* moduleName, plcbit* value);
-//bool subscribePLCOpen(plcstring* state, plcstring* moduleName, plcbit* command, AtnPlcOpenStatus *status );
-
-//void executeCommand( const STRING *action );
-unsigned short PLCOpenStatus( const STRING *action );
+unsigned short PLCOpenStatus( const STRING *command );
 
 bool forCommandGetPLCOpenStatus(plcstring* command, signed short index, unsigned short *status);